Switch to using Module.prepend instead of method aliasing in action_cable hook
